+++ AirplaneAPI/app/Http/Controllers/CourseController.php
@@ -19,30 +19,47 @@ class CourseController extends Controller {
     }
 
     public function deleteCourse(Request $request) {
-        $course = Course::find($request["id"]);
 
-        $course->delete();
-        return $course;
+        if(auth("sanctum")->user()->is_admin==1) {
 
+            $course = Course::find($request["id"]);
+
+            $course->delete();
+            return $course;
+
+        } else {
+                   
+            return $this->sendError("Sikertelen azonosítás", 401);
+        };
+        
     }
 
     public function updateCourse(Request $request) {
 
-        $course = Course::find($request["id"]);
-        $course -> course_name = $request["course_name"]; 
-        $course -> airplane_id = $request["airplane_id"];  
-        $course -> instructor = $request["instructor"];  
-        $course -> start_date = $request["start_date"];  
-        $course -> end_date = $request["end_date"]; 
-        $course -> course_fee = $request["course_fee"];  
+        if(auth("sanctum")->user()->is_admin==1) {
 
+            $course = Course::find($request["id"]);
+            $course -> course_name = $request["course_name"]; 
+            $course -> airplane_id = $request["airplane_id"];  
+            $course -> instructor = $request["instructor"];  
+            $course -> start_date = $request["start_date"];  
+            $course -> end_date = $request["end_date"]; 
+            $course -> course_fee = $request["course_fee"];  
 
+            $course -> save();
+            return $course;
 
-        $course -> save();
-        return $course;
+        } else {
+                   
+            return $this->sendError("Sikertelen azonosítás", 401);
+        };
+
+        
     }
 
     public function createCourse(Request $request) {
+
+        if(auth("sanctum")->user()->is_admin==1) {
 
         $course = new Course;
         $course -> id = $request["id"];
@@ -56,5 +73,10 @@ class CourseController extends Controller {
 
         $course -> save();
         return $course;
+
+    } else {
+                   
+        return $this->sendError("Sikertelen azonosítás", 401);
+    };
     }
 }
